Add `PECInsulator` boundary condition (#4943)
This PR adds a mixed PEC and insulator boundary condition. This allows
an insulator to be placed on a portion of the boundary. The rest of that
boundary will be PEC. Within the insulator portion, the tangential
fields can be specified on the boundary (as functions of space and
time). The normal fields and fields not specified are extrapolated to
the guard cells from the valid cells. The fields are specified in pairs,
the two tangential electric fields, and the two tangential magnetic
fields. In each pair, if one is set, the other will be zeroed if not
set.

A use case is the simulation of a dynamic pinch, driven by an external
current, represented as a time dependent B field on the boundary.


[PECinsulatorBC_warpX_summaryOnly.pdf](https://github.com/user-attachments/files/17637695/PECinsulatorBC_warpX_summaryOnly.pdf)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Edoardo Zoni <59625522+EZoni@users.noreply.github.com>