Implicit add filtering (#5086)
This adds filtering to the implicit solver, replacing PR #4600.

It is a simple change. All that is needed is adding a call to filter the
`Efield_fp` just before the particles are pushed. The current density is
already filtered in `SyncCurrentAndRho`.

The name of the routine `ApplyFilterJ` was changed to `ApplyFilterMF`
since it now has a more general usage.