Add strang implicit spectral em redo (#5027)
This implements use of the PSATD field advance coupled with the implicit
solver, using a Strang split advance.

- Advect Maxwell using PSATD with no J, ½ step
- Advance particles along with dE/dt = -J implicitly, full step,
iterating
- Advect Maxwell using PSATD with no J, ½ step

This requires the input parameter psatd.update_with_rho = 0.
With psatd.periodic_single_box_fft = 1, exact energy conservation is
obtained. Otherwise good conservation is seen, but not exact (will
depend on parameters).
Convergence is found for wpedt <= 1.9 (compared to wpedt < 0.25 for
FDTD).

This PR replaces PR #4662.

A task for a future PR would be to implement specialized source free
spectral advance routines (as noted in source comments).