Fix the time for the insulator boundary field evaluation (#5488)
In the application of the insulator boundary conditions, the field in
the insulator is specified using an expression that can be a function of
time. This PR updates the calls to that routine to pass in the correct
time, for example the time at the half step or the end of the step. This
requires passing the time down the call chain to the routines that apply
the insulator boundaries which makes this a fairly sizable set of
changes.

Note that in some places the variable names were changed to be more
precise about which time is passed in. For example, using `start_time`
whenever the time at the start of the step is expected.