Add Dan Barnes' effective potential (semi-implicit) Poisson solver (#5079)
This PR adds a new electrostatic solver based on the semi-implicit
scheme developed by [Barnes, Journal of Comp. Phys., 424 (2021),
109852](https://www.sciencedirect.com/science/article/pii/S0021999120306264?via%3Dihub)
(see Appendix A of reference).

The implementation was tested through the simulation of a Langmuir probe
placed inside a uniform plasma. To this end a 2d simulation was
performed with an initial uniform plasma with a conducting disk inserted
in the center of the domain. The disk is kept at 0 V while Neumann
boundary conditions are used for the domain boundary. Electrons and
hydrogen ions were injected from all the domain boundaries using the
`UniformFluxDistribution` in order to simulate the particle flux from a
uniform plasma. Thermal boundaries were used for the particles with the
same temperatures as the initial plasma particles. The expected outcome
of this simulation is that a sheath develops around the "Langmuir probe"
with value
$V_s = 0.5T_e\ln\left(2\pi\frac{m_e}{2m_p}(1 + \frac{T_i}{T_e})\right)
$.
A pre-sheath of roughly $0.7T_e$ is also expected to form but the exact
value of this depends on the domain size since the pre-sheath extends
for many Debye lengths.
The figure below shows an example outcome from a simulation as described
above in which the semi-implicit factor was set to $C_{SI} = 10$.

![image](https://github.com/user-attachments/assets/de4c4f44-3eb1-4e4b-8297-4c047d1c0d98)

Required before merging:

- [x] Merging of https://github.com/AMReX-Codes/amrex/pull/3968
- [x] Add example to CI tests
- [x] Update documentation

---------

Signed-off-by: roelof-groenewald <regroenewald@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Thomas Marks <marksta@umich.edu>