Fix FieldEnergy reduced diagnostic (#5498)
Fix FieldEnergy reduced diagnostic to account for partial cell volumes
on the domain boundaries. For fields that are nodal on the boundary, the
integral is only over the fraction of the cell centered about the field
value within the domain.

This PR also tidies up the code for the diagnostic.