DSMC: make the check of the number of processes more robust (#5515)
The previous code was brittle because it was not using
`AMREX_ALWAYS_ASSERT_WITH_MESSAGE` (which indeed cannot be used inside a
GPU kernel. In practice, a user could specify e.g. 10 scattering
processes and not get an error.

The new code checks the number of processes before calling the GPU
kernel, by using `AMREX_ALWAYS_ASSERT_WITH_MESSAGE`.

---------

Co-authored-by: Roelof Groenewald <40245517+roelof-groenewald@users.noreply.github.com>